requirements.txt
Flask==2.0.1
gunicorn==20.1.0
python-dotenv==0.16.0
discord.py==1.7.3
discord-webhook==0.14.0
dnspython==2.1.0
pymongo==3.12.1
certifi==2021.10.8
asyncio==3.4.3
squareup==16.0.0.20211117
